+++ R/package-maotai.R
@@ -6,9 +6,9 @@
 #' @name package-maotai
 #' @import Rdpack
 #' @importFrom cluster pam
-#' @importFrom stats as.dist knots ecdf rnorm runif quantile dist rgamma rgeom var
+#' @importFrom stats as.dist knots ecdf rnorm runif quantile dist rgamma rgeom var cov
 #' @importFrom shapes procGPA
-#' @importFrom mclust Mclust mclustBIC
+#' @importFrom Rtsne Rtsne
 #' @importFrom utils packageVersion
 #' @importFrom RSpectra eigs
 #' @importFrom Matrix rankMatrix