+++ src/RcppExports.cpp
@@ -6,6 +6,49 @@
 
 using namespace Rcpp;
 
+// compute_SSR
+double compute_SSR(arma::mat& D, arma::mat& Delta);
+RcppExport SEXP _maotai_compute_SSR(SEXP DSEXP, SEXP DeltaSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< arma::mat& >::type D(DSEXP);
+    Rcpp::traits::input_parameter< arma::mat& >::type Delta(DeltaSEXP);
+    rcpp_result_gen = Rcpp::wrap(compute_SSR(D, Delta));
+    return rcpp_result_gen;
+END_RCPP
+}
+// compute_stress
+double compute_stress(arma::mat& D, arma::mat& Dhat);
+RcppExport SEXP _maotai_compute_stress(SEXP DSEXP, SEXP DhatSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< arma::mat& >::type D(DSEXP);
+    Rcpp::traits::input_parameter< arma::mat& >::type Dhat(DhatSEXP);
+    rcpp_result_gen = Rcpp::wrap(compute_stress(D, Dhat));
+    return rcpp_result_gen;
+END_RCPP
+}
+// main_bmds
+Rcpp::List main_bmds(arma::mat D, arma::mat X0, double sigg0, double a, double alpha, int maxiter, double constant, bool verbose, arma::vec betas);
+RcppExport SEXP _maotai_main_bmds(SEXP DSEXP, SEXP X0SEXP, SEXP sigg0SEXP, SEXP aSEXP, SEXP alphaSEXP, SEXP maxiterSEXP, SEXP constantSEXP, SEXP verboseSEXP, SEXP betasSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< arma::mat >::type D(DSEXP);
+    Rcpp::traits::input_parameter< arma::mat >::type X0(X0SEXP);
+    Rcpp::traits::input_parameter< double >::type sigg0(sigg0SEXP);
+    Rcpp::traits::input_parameter< double >::type a(aSEXP);
+    Rcpp::traits::input_parameter< double >::type alpha(alphaSEXP);
+    Rcpp::traits::input_parameter< int >::type maxiter(maxiterSEXP);
+    Rcpp::traits::input_parameter< double >::type constant(constantSEXP);
+    Rcpp::traits::input_parameter< bool >::type verbose(verboseSEXP);
+    Rcpp::traits::input_parameter< arma::vec >::type betas(betasSEXP);
+    rcpp_result_gen = Rcpp::wrap(main_bmds(D, X0, sigg0, a, alpha, maxiter, constant, verbose, betas));
+    return rcpp_result_gen;
+END_RCPP
+}
 // aux_shortestpath
 Rcpp::NumericMatrix aux_shortestpath(NumericMatrix& wmat);
 RcppExport SEXP _maotai_aux_shortestpath(SEXP wmatSEXP) {
@@ -113,6 +156,17 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// emds_gamma0
+double emds_gamma0(arma::mat dmat);
+RcppExport SEXP _maotai_emds_gamma0(SEXP dmatSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< arma::mat >::type dmat(dmatSEXP);
+    rcpp_result_gen = Rcpp::wrap(emds_gamma0(dmat));
+    return rcpp_result_gen;
+END_RCPP
+}
 // cpp_pairwise_L2
 Rcpp::List cpp_pairwise_L2(arma::mat muA, arma::mat muB, arma::cube covA, arma::cube covB);
 RcppExport SEXP _maotai_cpp_pairwise_L2(SEXP muASEXP, SEXP muBSEXP, SEXP covASEXP, SEXP covBSEXP) {
@@ -136,6 +190,29 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< arma::vec& >::type tseq(tseqSEXP);
     Rcpp::traits::input_parameter< arma::vec& >::type fval(fvalSEXP);
     rcpp_result_gen = Rcpp::wrap(integrate_1d(tseq, fval));
+    return rcpp_result_gen;
+END_RCPP
+}
+// cpp_pdist
+arma::mat cpp_pdist(arma::mat X);
+RcppExport SEXP _maotai_cpp_pdist(SEXP XSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< arma::mat >::type X(XSEXP);
+    rcpp_result_gen = Rcpp::wrap(cpp_pdist(X));
+    return rcpp_result_gen;
+END_RCPP
+}
+// cpp_geigen
+Rcpp::List cpp_geigen(arma::mat& A, arma::mat& B);
+RcppExport SEXP _maotai_cpp_geigen(SEXP ASEXP, SEXP BSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< arma::mat& >::type A(ASEXP);
+    Rcpp::traits::input_parameter< arma::mat& >::type B(BSEXP);
+    rcpp_result_gen = Rcpp::wrap(cpp_geigen(A, B));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -195,6 +272,9 @@ END_RCPP
 }
 
 static const R_CallMethodDef CallEntries[] = {
+    {"_maotai_compute_SSR", (DL_FUNC) &_maotai_compute_SSR, 2},
+    {"_maotai_compute_stress", (DL_FUNC) &_maotai_compute_stress, 2},
+    {"_maotai_main_bmds", (DL_FUNC) &_maotai_main_bmds, 9},
     {"_maotai_aux_shortestpath", (DL_FUNC) &_maotai_aux_shortestpath, 1},
     {"_maotai_cppsub_2007Wang", (DL_FUNC) &_maotai_cppsub_2007Wang, 7},
     {"_maotai_gradF", (DL_FUNC) &_maotai_gradF, 3},
@@ -203,8 +283,11 @@ static const R_CallMethodDef CallEntries[] = {
     {"_maotai_solve_lyapunov", (DL_FUNC) &_maotai_solve_lyapunov, 3},
     {"_maotai_cpp_weiszfeld", (DL_FUNC) &_maotai_cpp_weiszfeld, 6},
     {"_maotai_cpp_kmeans", (DL_FUNC) &_maotai_cpp_kmeans, 2},
+    {"_maotai_emds_gamma0", (DL_FUNC) &_maotai_emds_gamma0, 1},
     {"_maotai_cpp_pairwise_L2", (DL_FUNC) &_maotai_cpp_pairwise_L2, 4},
     {"_maotai_integrate_1d", (DL_FUNC) &_maotai_integrate_1d, 2},
+    {"_maotai_cpp_pdist", (DL_FUNC) &_maotai_cpp_pdist, 1},
+    {"_maotai_cpp_geigen", (DL_FUNC) &_maotai_cpp_geigen, 2},
     {"_maotai_eval_gaussian", (DL_FUNC) &_maotai_eval_gaussian, 3},
     {"_maotai_eval_gaussian_data", (DL_FUNC) &_maotai_eval_gaussian_data, 3},
     {"_maotai_eval_gmm_data", (DL_FUNC) &_maotai_eval_gmm_data, 4},